Consolidate receiver into single console project
