--- src/Asynkron.OtelReceiver.Core/Asynkron.OtelReceiver.Core.csproj
@@ -1,57 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <RootNamespace>TraceLens</RootNamespace>
-        <LangVersion>preview</LangVersion>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <Protobuf Include="opentelemetry\proto\collector\metrics\v1\metrics_service.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-
-        <Protobuf Include="opentelemetry\proto\collector\logs\v1\logs_service.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="opentelemetry\proto\collector\trace\v1\trace_service.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="opentelemetry\proto\common\v1\common.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="opentelemetry\proto\logs\v1\logs.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="opentelemetry\proto\metrics\v1\metrics.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="opentelemetry\proto\resource\v1\resource.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="opentelemetry\proto\trace\v1\trace.proto">
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="tracelens.proto">
-
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <Protobuf Include="receiver_metrics.proto">
-            <Generator>MSBuild:Compile</Generator>
-            <GrpcServices>Both</GrpcServices>
-        </Protobuf>
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Google.Protobuf" Version="3.25.3" />
-        <PackageReference Include="Grpc.Net.Common" Version="2.61.0" />
-        <PackageReference Include="Grpc.Tools" Version="2.62.0">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Hocon.Json" Version="2.1.2" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
-    </ItemGroup>
-</Project>
