--- src/Asynkron.OtelReceiver.Host/Asynkron.OtelReceiver.Host.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Grpc.Net.Client" Version="2.61.0" />
-    <PackageReference Include="Spectre.Console" Version="0.47.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Asynkron.OtelReceiver.Core\Asynkron.OtelReceiver.Core.csproj" />
-  </ItemGroup>
-</Project>
