+++ src/Asynkron.OtelReceiver/TraceLens/ExtendedProtos/Metric.cs
