+++ src/Asynkron.OtelReceiver/TraceLens/Extensions.cs
