+++ src/Asynkron.OtelReceiver/TraceLens/Infra/ChannelExtensions.cs
