+++ src/Asynkron.OtelReceiver/TraceLens/Infra/Extensions.cs
