+++ src/Asynkron.OtelReceiver/TraceLens/Infra/JsonFlattener.cs
