+++ src/Asynkron.OtelReceiver/TraceLens/Infra/PlantUmlBuiltInStyles.cs
