+++ src/Asynkron.OtelReceiver/TraceLens/Infra/Translator.cs
