+++ src/Asynkron.OtelReceiver/TraceLens/Model/Attributes.cs
