+++ src/Asynkron.OtelReceiver/TraceLens/Model/Call.cs
