+++ src/Asynkron.OtelReceiver/TraceLens/Model/CallKind.cs
