+++ src/Asynkron.OtelReceiver/TraceLens/Model/Component.cs
