+++ src/Asynkron.OtelReceiver/TraceLens/Model/ComponentKind.cs
