+++ src/Asynkron.OtelReceiver/TraceLens/Model/ComponentMetadata.cs
