+++ src/Asynkron.OtelReceiver/TraceLens/Model/ComponentModel.cs
