+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/AzureExtractor.cs
