+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/DbExtractor.cs
