+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/DbStatementExtractor.cs
