+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/ExternalHttpEndpointExtractor.cs
