+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/HttpEndpointExtractor.cs
