+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/IExtractor.cs
