+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/ProtoActorExtractor.cs
