+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/QueueExtractor.cs
