+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/RootExtractor.cs
