+++ src/Asynkron.OtelReceiver/TraceLens/Model/Extractors/TemporalExtractor.cs
