+++ src/Asynkron.OtelReceiver/TraceLens/Model/Group.cs
