+++ src/Asynkron.OtelReceiver/TraceLens/Model/LogEntry.cs
