+++ src/Asynkron.OtelReceiver/TraceLens/Model/Metric.cs
