+++ src/Asynkron.OtelReceiver/TraceLens/Model/ServiceModel.cs
