+++ src/Asynkron.OtelReceiver/TraceLens/Model/Span.cs
