+++ src/Asynkron.OtelReceiver/TraceLens/Model/SpanDescription.cs
