+++ src/Asynkron.OtelReceiver/TraceLens/Model/Strings.cs
