+++ src/Asynkron.OtelReceiver/TraceLens/Model/TraceLensModel.cs
