+++ src/Asynkron.OtelReceiver/TraceLens/StringUtils.cs
