+++ src/Asynkron.OtelReceiver/opentelemetry/proto/collector/metrics/v1/metrics_service_http.yaml
