+++ src/Asynkron.OtelReceiver/opentelemetry/proto/collector/trace/v1/trace_service_http.yaml
