+++ src/Asynkron.OtelReceiver/opentelemetry/proto/metrics/v1/metrics.proto
