+++ src/Asynkron.OtelReceiver/opentelemetry/proto/resource/v1/resource.proto
