+++ src/Asynkron.OtelReceiver/opentelemetry/proto/trace/v1/trace.proto
