+++ src/Asynkron.OtelReceiver/receiver_metrics.proto
